refactor: split off logging
