+++ packages/node/src/core/getLatestPackageTags.ts
@@ -1,7 +1,8 @@
 import { Workspace } from '@yarnpkg/core'
 import * as pluginNPM from '@yarnpkg/plugin-npm'
 
-import logging from '../logging'
+import logging from 'monodeploy-logging'
+
 import type {
     MonodeployConfiguration,
     PackageTagMap,
