+++ packages/node/src/core/prependChangelogFile.ts
@@ -1,7 +1,8 @@
 import { promises as fs } from 'fs'
 import path from 'path'
 
-import logging from '../logging'
+import logging from 'monodeploy-logging'
+
 import type {
     ChangesetSchema,
     MonodeployConfiguration,
