+++ packages/node/src/monodeploy.ts
@@ -4,6 +4,8 @@ import { getPluginConfiguration } from '@yarnpkg/cli'
 import { Cache, Configuration, Project, ThrowReport } from '@yarnpkg/core'
 import { PortablePath } from '@yarnpkg/fslib'
 
+import logging from 'monodeploy-logging'
+
 import applyReleases from './core/applyReleases'
 import getExplicitVersionStrategies from './core/getExplicitVersionStrategies'
 import getImplicitVersionStrategies from './core/getImplicitVersionStrategies'
@@ -11,7 +13,6 @@ import getLatestPackageTags from './core/getLatestPackageTags'
 import prependChangelogFile from './core/prependChangelogFile'
 import publishPackages from './core/publishPackages'
 import writeChangesetFile from './core/writeChangesetFile'
-import logging from './logging'
 import type {
     ChangesetSchema,
     MonodeployConfiguration,
